+++ src/main.js
@@ -5,7 +5,7 @@ import StatsView from './view/stats.js';
 import FilmCounterView from './view/film-counter.js';
 import FilterMenuPresenter from './presenter/filter-menu.js';
 import FilmBoardPresenter from './presenter/film-board.js';
-import { remove, render } from './utils.js';
+import { remove, render } from './utils/utils.js';
 import { RenderPosition, PageState, UpdateType } from './consts.js';
 import Api from './api/api.js';
 import Store from './api/store.js';
