+++ src/presenter/film-popup.js
@@ -1,7 +1,7 @@
 ﻿import FilmPopupView from '../view/film-popup.js';
 import PopupCommentView from '../view/popup-comment.js';
 import PopupControlsView from '../view/popup-controls.js';
-import { render, remove, shake } from '../utils.js';
+import { render, remove, shake } from '../utils/utils.js';
 import { RenderPosition, UpdateType, PopupState } from '../consts.js';
 
 export default class FilmPopup {
