+++ src/utils/utils.js
@@ -1,30 +1,10 @@
-﻿import { SortType, FilterType, UserProfileRatings, FilterPeriod } from './consts.js';
-import Abstract from './view/abstract.js';
+﻿import { UserProfileRatings } from '../consts.js';
+import Abstract from '../view/abstract.js';
 import dayjs from 'dayjs';
 import duration from 'dayjs/plugin/duration';
 import relativeTime from 'dayjs/plugin/relativeTime';
 dayjs.extend(duration);
 dayjs.extend(relativeTime);
-
-export const PeriodStrategy = {
-  [FilterPeriod.TODAY]: () => dayjs().subtract(1, 'day').toDate(),
-  [FilterPeriod.WEEK]: () => dayjs().subtract(1, 'week').toDate(),
-  [FilterPeriod.MONTH]: () => dayjs().subtract(1, 'month').toDate(),
-  [FilterPeriod.YEAR]: () => dayjs().subtract(1, 'year').toDate(),
-};
-
-export const FilterStrategy = {
-  [FilterType.ALL]: (films) => films,
-  [FilterType.WATCHLIST]: (films) => films.filter((film) => film.userDetails.isInWatchlist),
-  [FilterType.HISTORY]: (films) => films.filter((film) => film.userDetails.isWatched),
-  [FilterType.FAVORITES]: (films) => films.filter((film) => film.userDetails.isFavorite),
-};
-
-export const SortStrategy = {
-  [SortType.BY_RATING]: (a, b) => b.filmInfo.rating - a.filmInfo.rating,
-  [SortType.BY_COMMENT_AMOUNT]: (a, b) => b.comments.length - a.comments.length,
-  [SortType.BY_DATE]: (a, b) => dayjs(b.filmInfo.releaseDate).diff(a.filmInfo.releaseDate),
-};
 
 export const render = (container, element, place) => {
   if (container instanceof Abstract) {
